fincept-terminal-desktop/src-tauri/resources/requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.2
aiosignal==1.4.0
attrs==25.4.0
beautifulsoup4==4.14.2
certifi==2025.11.12
cffi==2.0.0
charset-normalizer==3.4.4
curl_cffi==0.13.0
databento==0.66.0
databento-dbn==0.44.0
dnspython==2.8.0
feedparser==6.0.12
frozendict==2.4.7
frozenlist==1.8.0
gnews==0.4.2
idna==3.11
multidict==6.7.0
multitasking==0.0.12
numpy>=1.26.4,<3.0
pandas>=2.3.3,<3.0
vectorbt==0.28.2
backtesting==0.6.5
# fast-trade==1.0.2  # Commented out - not tested with Python 3.12, use backtesting.py instead
peewee==3.18.3
pip_system_certs==5.3
platformdirs==4.5.0
propcache==0.4.1
protobuf>=3.20.2,<6.0.0
pyarrow==22.0.0
pycparser==2.23
python-dateutil==2.9.0.post0
pytz==2025.2
requests==2.32.5
sgmllib3k==1.0.0
six==1.17.0
soupsieve==2.8
ta==0.11.0
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
websockets==15.0.1
yarl==1.22.0
yfinance==0.2.66
zstandard==0.25.0
jinja2==3.1.3
weasyprint==62.3
matplotlib>=3.9.3,<4.0
# LLM Provider Packages (for finagent_core)
openai>=2.8.1,<3.0
anthropic>=0.71.0,<1.0
google-generativeai>=0.8.5,<1.0

# AI Quant Lab - Qlib and RD-Agent
pyqlib==0.9.7
rdagent==0.8.0

# Qlib core dependencies
cvxpy>=1.0.0
dill>=0.3.0
fire>=0.4.0
gym>=0.26.0
jupyter>=1.0.0
loguru>=0.6.0
mlflow>=2.0.0
nbconvert>=7.0.0
pydantic-settings>=2.0.0
python-redis-lock>=4.0.0
ruamel.yaml>=0.17.38

# RD-Agent core dependencies
langchain>=0.1.0
langchain-community>=0.0.1
litellm>=1.73
nbformat>=5.0.0
pandarallel>=1.6.0
rich>=13.0.0
seaborn>=0.12.0
selenium>=4.0.0
setuptools-scm>=8.0.0
streamlit>=1.47
tiktoken>=0.5.0
typer>=0.9.0
tables>=3.8.0

# ML/DL libraries for Qlib
lightgbm>=4.0.0
xgboost>=2.0.0
catboost>=1.2.0
torch>=2.0.0
scikit-learn>=1.3.0
scipy>=1.10.0

# EdgarTools - SEC EDGAR data library
edgartools>=5.6.0

# HDX - Humanitarian Data Exchange
# hdx-python-api>=6.3.3  # Removed - using direct CKAN API instead (no dependencies required)

# Optional RD-Agent dependencies (Azure, Kaggle)
# Uncomment if needed:
# azure-ai-formrecognizer
# azure-ai-inference
# azure-identity
# azureml-mlflow
# humanize
# kaggle
# prefect
# pydantic-ai-slim
# python-levenshtein
# randomname
# st-theme
# webdriver-manager
financepy==1.0.1
ffn==1.1.2
plotly==6.5.0
gs-quant==1.4.67
rateslib==2.1.1
fortitudo.tech==1.2
cvxopt==1.3.2
FinQuant==0.7.0
vnpy==4.3.0
talipp==2.7.0
py_vollib==1.0.1
tsmoothie==1.0.5
gluonts[torch]==0.16.2
statsmodels==0.14.6
pypme==0.6.2
pmdarima==2.1.1
functime==0.1.10
